#80 - fix: Added coordinates to location form

+++ backend/src/main/java/com/eventus/backend/models/Coordinates.java
@@ -7,10 +7,10 @@ import javax.validation.constraints.NotBlank;
 public class Coordinates {
 
     @NotBlank
-    private Long latitude;
+    private Double latitude;
 
     @NotBlank
-    private Long longitude;
+    private Double longitude;
 
 
     public Coordinates(){
@@ -18,23 +18,23 @@ public class Coordinates {
     }
     
     public Coordinates(String latitude, String longitude) {
-        this.latitude = Long.valueOf(latitude);
-        this.longitude = Long.valueOf(longitude);
+        this.latitude = Double.valueOf(latitude);
+        this.longitude = Double.valueOf(longitude);
     }
 
-    public Long getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Long latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Long getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Long longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 }
